Update model test cases
Trigger CI
