+++ tests/test_models.py
@@ -4,29 +4,19 @@
 
 import importlib
 import os
-import tempfile
-import traceback
 import unittest
-from typing import Any, ClassVar, Mapping, Optional, Type
-from unittest.mock import patch
+from typing import Optional
 
 import numpy
-import pytest
 import torch
-from click.testing import CliRunner, Result
-from torch import nn, optim
-from torch.optim import SGD
-from torch.optim.adagrad import Adagrad
+from torch import nn
 
 import pykeen.experiments
 import pykeen.models
-from pykeen.datasets.kinships import KINSHIPS_TRAIN_PATH
-from pykeen.datasets.nations import NATIONS_TEST_PATH, NATIONS_TRAIN_PATH, Nations
 from pykeen.models import (
     EntityEmbeddingModel, EntityRelationEmbeddingModel, Model, MultimodalModel, _MODELS,
     _OldAbstractModel,
 )
-from pykeen.models.cli import build_cli_from_cls
 from pykeen.models.predict import get_novelty_mask, predict
 from pykeen.models.unimodal.rgcn import (
     inverse_indegree_edge_weights,
@@ -35,9 +25,8 @@ from pykeen.models.unimodal.rgcn import (
 )
 from pykeen.models.unimodal.trans_d import _project_entity
 from pykeen.nn import Embedding, RepresentationModule
-from pykeen.training import LCWATrainingLoop, SLCWATrainingLoop, TrainingLoop
-from pykeen.triples import TriplesFactory
-from pykeen.utils import all_in_bounds, clamp_norm, extend_batch, set_random_seed
+from pykeen.utils import all_in_bounds, clamp_norm, extend_batch
+from tests import cases
 
 SKIP_MODULES = {
     Model.__name__,
@@ -80,475 +69,13 @@ class _CustomRepresentations(RepresentationModule):
         return x.unsqueeze(dim=1)
 
 
-class _ModelTestCase:
-    """A test case for quickly defining common tests for KGE models."""
-
-    #: The class of the model to test
-    model_cls: ClassVar[Type[Model]]
-
-    #: Additional arguments passed to the model's constructor method
-    model_kwargs: ClassVar[Optional[Mapping[str, Any]]] = None
-
-    #: Additional arguments passed to the training loop's constructor method
-    training_loop_kwargs: ClassVar[Optional[Mapping[str, Any]]] = None
-
-    #: The triples factory instance
-    factory: TriplesFactory
-
-    #: The model instance
-    model: Model
-
-    #: The batch size for use for forward_* tests
-    batch_size: int = 20
-
-    #: The embedding dimensionality
-    embedding_dim: int = 3
-
-    #: Whether to create inverse triples (needed e.g. by ConvE)
-    create_inverse_triples: bool = False
-
-    #: The sampler to use for sLCWA (different e.g. for R-GCN)
-    sampler = 'default'
-
-    #: The batch size for use when testing training procedures
-    train_batch_size = 400
-
-    #: The number of epochs to train the model
-    train_num_epochs = 2
-
-    #: A random number generator from torch
-    generator: torch.Generator
-
-    #: The number of parameters which receive a constant (i.e. non-randomized)
-    # initialization
-    num_constant_init: int = 0
-
-    def setUp(self) -> None:
-        """Set up the test case with a triples factory and model."""
-        _, self.generator, _ = set_random_seed(42)
-
-        dataset = Nations(create_inverse_triples=self.create_inverse_triples)
-        self.factory = dataset.training
-        self.model = self.model_cls(
-            self.factory,
-            embedding_dim=self.embedding_dim,
-            **(self.model_kwargs or {}),
-        ).to_device_()
-
-    def test_get_grad_parameters(self):
-        """Test the model's ``get_grad_params()`` method."""
-        # assert there is at least one trainable parameter
-        assert len(list(self.model.get_grad_params())) > 0
-
-        # Check that all the parameters actually require a gradient
-        for parameter in self.model.get_grad_params():
-            assert parameter.requires_grad
-
-        # Try to initialize an optimizer
-        optimizer = SGD(params=self.model.get_grad_params(), lr=1.0)
-        assert optimizer is not None
-
-    def test_reset_parameters_(self):
-        """Test :func:`Model.reset_parameters_`."""
-        # get model parameters
-        params = list(self.model.parameters())
-        old_content = {
-            id(p): p.data.detach().clone()
-            for p in params
-        }
-
-        # re-initialize
-        self.model.reset_parameters_()
-
-        # check that the operation works in-place
-        new_params = list(self.model.parameters())
-        assert set(id(np) for np in new_params) == set(id(p) for p in params)
-
-        # check that the parameters where modified
-        num_equal_weights_after_re_init = sum(
-            1
-            for np in new_params
-            if (np.data == old_content[id(np)]).all()
-        )
-        assert num_equal_weights_after_re_init == self.num_constant_init, (
-            num_equal_weights_after_re_init, self.num_constant_init,
-        )
-
-    def _check_scores(self, batch, scores) -> None:
-        """Check the scores produced by a forward function."""
-        # check for finite values by default
-        self.assertTrue(torch.all(torch.isfinite(scores)).item(), f'Some scores were not finite:\n{scores}')
-
-        # check whether a gradient can be back-propgated
-        scores.mean().backward()
-
-    def test_save(self) -> None:
-        """Test that the model can be saved properly."""
-        with tempfile.TemporaryDirectory() as temp_directory:
-            torch.save(self.model, os.path.join(temp_directory, 'model.pickle'))
-
-    def test_score_hrt(self) -> None:
-        """Test the model's ``score_hrt()`` function."""
-        batch = self.factory.mapped_triples[:self.batch_size, :].to(self.model.device)
-        try:
-            scores = self.model.score_hrt(batch)
-        except RuntimeError as e:
-            if str(e) == 'fft: ATen not compiled with MKL support':
-                self.skipTest(str(e))
-            else:
-                raise e
-        assert scores.shape == (self.batch_size, 1)
-        self._check_scores(batch, scores)
-
-    def test_score_t(self) -> None:
-        """Test the model's ``score_t()`` function."""
-        batch = self.factory.mapped_triples[:self.batch_size, :2].to(self.model.device)
-        # assert batch comprises (head, relation) pairs
-        assert batch.shape == (self.batch_size, 2)
-        assert (batch[:, 0] < self.factory.num_entities).all()
-        assert (batch[:, 1] < self.factory.num_relations).all()
-        try:
-            scores = self.model.score_t(batch)
-        except NotImplementedError:
-            self.fail(msg='Score_o not yet implemented')
-        except RuntimeError as e:
-            if str(e) == 'fft: ATen not compiled with MKL support':
-                self.skipTest(str(e))
-            else:
-                raise e
-        assert scores.shape == (self.batch_size, self.model.num_entities)
-        self._check_scores(batch, scores)
-
-    def test_score_h(self) -> None:
-        """Test the model's ``score_h()`` function."""
-        batch = self.factory.mapped_triples[:self.batch_size, 1:].to(self.model.device)
-        # assert batch comprises (relation, tail) pairs
-        assert batch.shape == (self.batch_size, 2)
-        assert (batch[:, 0] < self.factory.num_relations).all()
-        assert (batch[:, 1] < self.factory.num_entities).all()
-        try:
-            scores = self.model.score_h(batch)
-        except NotImplementedError:
-            self.fail(msg='Score_s not yet implemented')
-        except RuntimeError as e:
-            if str(e) == 'fft: ATen not compiled with MKL support':
-                self.skipTest(str(e))
-            else:
-                raise e
-        assert scores.shape == (self.batch_size, self.model.num_entities)
-        self._check_scores(batch, scores)
-
-    @pytest.mark.slow
-    def test_train_slcwa(self) -> None:
-        """Test that sLCWA training does not fail."""
-        loop = SLCWATrainingLoop(
-            model=self.model,
-            optimizer=Adagrad(params=self.model.get_grad_params(), lr=0.001),
-            **(self.training_loop_kwargs or {}),
-        )
-        losses = self._safe_train_loop(
-            loop,
-            num_epochs=self.train_num_epochs,
-            batch_size=self.train_batch_size,
-            sampler=self.sampler,
-        )
-        self.assertIsInstance(losses, list)
-
-    @pytest.mark.slow
-    def test_train_lcwa(self) -> None:
-        """Test that LCWA training does not fail."""
-        loop = LCWATrainingLoop(
-            model=self.model,
-            optimizer=Adagrad(params=self.model.get_grad_params(), lr=0.001),
-            **(self.training_loop_kwargs or {}),
-        )
-        losses = self._safe_train_loop(
-            loop,
-            num_epochs=self.train_num_epochs,
-            batch_size=self.train_batch_size,
-            sampler='default',
-        )
-        self.assertIsInstance(losses, list)
-
-    def _safe_train_loop(self, loop: TrainingLoop, num_epochs, batch_size, sampler):
-        try:
-            losses = loop.train(num_epochs=num_epochs, batch_size=batch_size, sampler=sampler, use_tqdm=False)
-        except RuntimeError as e:
-            if str(e) == 'fft: ATen not compiled with MKL support':
-                self.skipTest(str(e))
-            else:
-                raise e
-        else:
-            return losses
-
-    def test_save_load_model_state(self):
-        """Test whether a saved model state can be re-loaded."""
-        original_model = self.model_cls(
-            self.factory,
-            embedding_dim=self.embedding_dim,
-            random_seed=42,
-            **(self.model_kwargs or {}),
-        ).to_device_()
-
-        loaded_model = self.model_cls(
-            self.factory,
-            embedding_dim=self.embedding_dim,
-            random_seed=21,
-            **(self.model_kwargs or {}),
-        ).to_device_()
-
-        def _equal_embeddings(a: RepresentationModule, b: RepresentationModule) -> bool:
-            """Test whether two embeddings are equal."""
-            return (a(indices=None) == b(indices=None)).all()
-
-        if isinstance(original_model, EntityEmbeddingModel):
-            assert not _equal_embeddings(original_model.entity_embeddings, loaded_model.entity_embeddings)
-        if isinstance(original_model, EntityRelationEmbeddingModel):
-            assert not _equal_embeddings(original_model.relation_embeddings, loaded_model.relation_embeddings)
-
-        with tempfile.TemporaryDirectory() as tmpdirname:
-            file_path = os.path.join(tmpdirname, 'test.pt')
-            original_model.save_state(path=file_path)
-            loaded_model.load_state(path=file_path)
-        if isinstance(original_model, EntityEmbeddingModel):
-            assert _equal_embeddings(original_model.entity_embeddings, loaded_model.entity_embeddings)
-        if isinstance(original_model, EntityRelationEmbeddingModel):
-            assert _equal_embeddings(original_model.relation_embeddings, loaded_model.relation_embeddings)
-
-    @property
-    def cli_extras(self):
-        kwargs = self.model_kwargs or {}
-        extras = [
-            '--silent',
-        ]
-        for k, v in kwargs.items():
-            extras.append('--' + k.replace('_', '-'))
-            extras.append(str(v))
-
-        # For the high/low memory test cases of NTN, SE, etc.
-        if self.training_loop_kwargs and 'automatic_memory_optimization' in self.training_loop_kwargs:
-            automatic_memory_optimization = self.training_loop_kwargs.get('automatic_memory_optimization')
-            if automatic_memory_optimization is True:
-                extras.append('--automatic-memory-optimization')
-            elif automatic_memory_optimization is False:
-                extras.append('--no-automatic-memory-optimization')
-            # else, leave to default
-
-        extras += [
-            '--number-epochs', self.train_num_epochs,
-            '--embedding-dim', self.embedding_dim,
-            '--batch-size', self.train_batch_size,
-        ]
-        extras = [str(e) for e in extras]
-        return extras
-
-    @pytest.mark.slow
-    def test_cli_training_nations(self):
-        """Test running the pipeline on almost all models with only training data."""
-        self._help_test_cli(['-t', NATIONS_TRAIN_PATH] + self.cli_extras)
-
-    @pytest.mark.slow
-    def test_cli_training_kinships(self):
-        """Test running the pipeline on almost all models with only training data."""
-        self._help_test_cli(['-t', KINSHIPS_TRAIN_PATH] + self.cli_extras)
-
-    @pytest.mark.slow
-    def test_cli_training_nations_testing(self):
-        """Test running the pipeline on almost all models with only training data."""
-        self._help_test_cli(['-t', NATIONS_TRAIN_PATH, '-q', NATIONS_TEST_PATH] + self.cli_extras)
-
-    def _help_test_cli(self, args):
-        """Test running the pipeline on all models."""
-        if issubclass(self.model_cls, pykeen.models.RGCN):
-            self.skipTest('There is a problem with non-reproducible unittest for R-GCN.')
-        runner = CliRunner()
-        cli = build_cli_from_cls(self.model_cls)
-        # TODO: Catch HolE MKL error?
-        result: Result = runner.invoke(cli, args)
-
-        self.assertEqual(
-            0,
-            result.exit_code,
-            msg=f'''
-Command
-=======
-$ pykeen train {self.model_cls.__name__.lower()} {' '.join(args)}
-
-Output
-======
-{result.output}
-
-Exception
-=========
-{result.exc_info[1]}
-
-Traceback
-=========
-{''.join(traceback.format_tb(result.exc_info[2]))}
-            ''',
-        )
-
-    def test_has_hpo_defaults(self):
-        """Test that there are defaults for HPO."""
-        try:
-            d = self.model_cls.hpo_default
-        except AttributeError:
-            self.fail(msg=f'{self.model_cls.__name__} is missing hpo_default class attribute')
-        else:
-            self.assertIsInstance(d, dict)
-
-    def test_post_parameter_update_regularizer(self):
-        """Test whether post_parameter_update resets the regularization term."""
-        # set regularizer term
-        self.model.regularizer.regularization_term = None
-
-        # call post_parameter_update
-        self.model.post_parameter_update()
-
-        # assert that the regularization term has been reset
-        assert self.model.regularizer.regularization_term == torch.zeros(1, dtype=torch.float, device=self.model.device)
-
-    def test_post_parameter_update(self):
-        """Test whether post_parameter_update correctly enforces model constraints."""
-        # do one optimization step
-        opt = optim.SGD(params=self.model.parameters(), lr=1.)
-        batch = self.factory.mapped_triples[:self.batch_size, :].to(self.model.device)
-        scores = self.model.score_hrt(hrt_batch=batch)
-        fake_loss = scores.mean()
-        fake_loss.backward()
-        opt.step()
-
-        # call post_parameter_update
-        self.model.post_parameter_update()
-
-        # check model constraints
-        self._check_constraints()
-
-    def _check_constraints(self):
-        """Check model constraints."""
-
-    def test_score_h_with_score_hrt_equality(self) -> None:
-        """Test the equality of the model's  ``score_h()`` and ``score_hrt()`` function."""
-        batch = self.factory.mapped_triples[:self.batch_size, 1:].to(self.model.device)
-        self.model.eval()
-        # assert batch comprises (relation, tail) pairs
-        assert batch.shape == (self.batch_size, 2)
-        assert (batch[:, 0] < self.factory.num_relations).all()
-        assert (batch[:, 1] < self.factory.num_entities).all()
-        try:
-            scores_h = self.model.score_h(batch)
-            scores_hrt = super(self.model.__class__, self.model).score_h(batch)
-        except NotImplementedError:
-            self.fail(msg='Score_h not yet implemented')
-        except RuntimeError as e:
-            if str(e) == 'fft: ATen not compiled with MKL support':
-                self.skipTest(str(e))
-            else:
-                raise e
-
-        assert torch.allclose(scores_h, scores_hrt, atol=1e-06)
-
-    def test_score_r_with_score_hrt_equality(self) -> None:
-        """Test the equality of the model's  ``score_r()`` and ``score_hrt()`` function."""
-        batch = self.factory.mapped_triples[:self.batch_size, [0, 2]].to(self.model.device)
-        self.model.eval()
-        # assert batch comprises (relation, tail) pairs
-        assert batch.shape == (self.batch_size, 2)
-        assert (batch[:, 0] < self.factory.num_entities).all()
-        assert (batch[:, 1] < self.factory.num_entities).all()
-        try:
-            scores_r = self.model.score_r(batch)
-            scores_hrt = super(self.model.__class__, self.model).score_r(batch)
-        except NotImplementedError:
-            self.fail(msg='Score_h not yet implemented')
-        except RuntimeError as e:
-            if str(e) == 'fft: ATen not compiled with MKL support':
-                self.skipTest(str(e))
-            else:
-                raise e
-
-        assert torch.allclose(scores_r, scores_hrt, atol=1e-06)
-
-    def test_score_t_with_score_hrt_equality(self) -> None:
-        """Test the equality of the model's  ``score_t()`` and ``score_hrt()`` function."""
-        batch = self.factory.mapped_triples[:self.batch_size, :-1].to(self.model.device)
-        self.model.eval()
-        # assert batch comprises (relation, tail) pairs
-        assert batch.shape == (self.batch_size, 2)
-        assert (batch[:, 0] < self.factory.num_entities).all()
-        assert (batch[:, 1] < self.factory.num_relations).all()
-        try:
-            scores_t = self.model.score_t(batch)
-            scores_hrt = super(self.model.__class__, self.model).score_t(batch)
-        except NotImplementedError:
-            self.fail(msg='Score_h not yet implemented')
-        except RuntimeError as e:
-            if str(e) == 'fft: ATen not compiled with MKL support':
-                self.skipTest(str(e))
-            else:
-                raise e
-
-        assert torch.allclose(scores_t, scores_hrt, atol=1e-06)
-
-    def test_reset_parameters_constructor_call(self):
-        """Tests whether reset_parameters is called in the constructor."""
-        with patch.object(self.model_cls, 'reset_parameters_', return_value=None) as mock_method:
-            try:
-                self.model_cls(
-                    triples_factory=self.factory,
-                    embedding_dim=self.embedding_dim,
-                    **(self.model_kwargs or {}),
-                )
-            except TypeError as error:
-                assert error.args == ("'NoneType' object is not callable",)
-            mock_method.assert_called_once()
-
-    def test_custom_representations(self):
-        """Tests whether we can provide custom representations."""
-        if isinstance(self.model, EntityEmbeddingModel):
-            old_embeddings = self.model.entity_embeddings
-            self.model.entity_embeddings = _CustomRepresentations(
-                num_entities=self.factory.num_entities,
-                embedding_dim=old_embeddings.embedding_dim,
-            )
-            # call some functions
-            self.model.reset_parameters_()
-            self.test_score_hrt()
-            self.test_score_t()
-            # reset to old state
-            self.model.entity_embeddings = old_embeddings
-        elif isinstance(self.model, EntityRelationEmbeddingModel):
-            old_embeddings = self.model.relation_embeddings
-            self.model.relation_embeddings = _CustomRepresentations(
-                num_entities=self.factory.num_relations,
-                embedding_dim=old_embeddings.embedding_dim,
-            )
-            # call some functions
-            self.model.reset_parameters_()
-            self.test_score_hrt()
-            self.test_score_t()
-            # reset to old state
-            self.model.relation_embeddings = old_embeddings
-        else:
-            self.skipTest(f'Not testing custom representations for model: {self.model.__class__.__name__}')
-
-
-class _DistanceModelTestCase(_ModelTestCase):
-    """A test case for distance-based models."""
-
-    def _check_scores(self, batch, scores) -> None:
-        super()._check_scores(batch=batch, scores=scores)
-        # Distance-based model
-        assert (scores <= 0.0).all()
-
-
-class TestComplex(_ModelTestCase, unittest.TestCase):
+class TestComplex(cases.ModelTestCase):
     """Test the ComplEx model."""
 
     model_cls = pykeen.models.ComplEx
 
 
-class TestConvE(_ModelTestCase, unittest.TestCase):
+class TestConvE(cases.ModelTestCase):
     """Test the ConvE model."""
 
     model_cls = pykeen.models.ConvE
@@ -566,7 +93,7 @@ class TestConvE(_ModelTestCase, unittest.TestCase):
     num_constant_init = 7
 
 
-class TestConvKB(_ModelTestCase, unittest.TestCase):
+class TestConvKB(cases.ModelTestCase):
     """Test the ConvKB model."""
 
     model_cls = pykeen.models.ConvKB
@@ -577,7 +104,7 @@ class TestConvKB(_ModelTestCase, unittest.TestCase):
     num_constant_init = 3
 
 
-class TestDistMult(_ModelTestCase, unittest.TestCase):
+class TestDistMult(cases.ModelTestCase):
     """Test the DistMult model."""
 
     model_cls = pykeen.models.DistMult
@@ -637,7 +164,7 @@ class TestDistMult(_ModelTestCase, unittest.TestCase):
         self._test_score_all_triples(k=None)
 
 
-class TestERMLP(_ModelTestCase, unittest.TestCase):
+class TestERMLP(cases.ModelTestCase):
     """Test the ERMLP model."""
 
     model_cls = pykeen.models.ERMLP
@@ -648,7 +175,7 @@ class TestERMLP(_ModelTestCase, unittest.TestCase):
     num_constant_init = 2
 
 
-class TestERMLPE(_ModelTestCase, unittest.TestCase):
+class TestERMLPE(cases.ModelTestCase):
     """Test the extended ERMLP model."""
 
     model_cls = pykeen.models.ERMLPE
@@ -659,7 +186,7 @@ class TestERMLPE(_ModelTestCase, unittest.TestCase):
     num_constant_init = 4
 
 
-class TestHolE(_ModelTestCase, unittest.TestCase):
+class TestHolE(cases.ModelTestCase):
     """Test the HolE model."""
 
     model_cls = pykeen.models.HolE
@@ -672,24 +199,7 @@ class TestHolE(_ModelTestCase, unittest.TestCase):
         assert all_in_bounds(self.model.entity_embeddings(indices=None).norm(p=2, dim=-1), high=1., a_tol=_EPSILON)
 
 
-class _TestKG2E(_ModelTestCase):
-    """General tests for the KG2E model."""
-
-    model_cls = pykeen.models.KG2E
-
-    def _check_constraints(self):
-        """Check model constraints.
-
-        * Entity and relation embeddings have to have at most unit L2 norm.
-        * Covariances have to have values between c_min and c_max
-        """
-        for embedding in (self.model.entity_embeddings, self.model.relation_embeddings):
-            assert all_in_bounds(embedding(indices=None).norm(p=2, dim=-1), high=1., a_tol=_EPSILON)
-        for cov in (self.model.entity_covariances, self.model.relation_covariances):
-            assert all_in_bounds(cov(indices=None), low=self.model.c_min, high=self.model.c_max)
-
-
-class TestKG2EWithKL(_TestKG2E, unittest.TestCase):
+class TestKG2EWithKL(cases.BaseKG2ETest):
     """Test the KG2E model with KL similarity."""
 
     model_kwargs = {
@@ -697,7 +207,7 @@ class TestKG2EWithKL(_TestKG2E, unittest.TestCase):
     }
 
 
-class TestKG2EWithEL(_TestKG2E, unittest.TestCase):
+class TestKG2EWithEL(cases.BaseKG2ETest):
     """Test the KG2E model with EL similarity."""
 
     model_kwargs = {
@@ -705,19 +215,7 @@ class TestKG2EWithEL(_TestKG2E, unittest.TestCase):
     }
 
 
-class _BaseNTNTest(_ModelTestCase, unittest.TestCase):
-    """Test the NTN model."""
-
-    model_cls = pykeen.models.NTN
-
-    def test_can_slice(self):
-        """Test that the slicing properties are calculated correctly."""
-        self.assertTrue(self.model.can_slice_h)
-        self.assertFalse(self.model.can_slice_r)
-        self.assertTrue(self.model.can_slice_t)
-
-
-class TestNTNLowMemory(_BaseNTNTest):
+class TestNTNLowMemory(cases.BaseNTNTest):
     """Test the NTN model with automatic memory optimization."""
 
     model_kwargs = {
@@ -729,7 +227,7 @@ class TestNTNLowMemory(_BaseNTNTest):
     }
 
 
-class TestNTNHighMemory(_BaseNTNTest):
+class TestNTNHighMemory(cases.BaseNTNTest):
     """Test the NTN model without automatic memory optimization."""
 
     model_kwargs = {
@@ -741,33 +239,19 @@ class TestNTNHighMemory(_BaseNTNTest):
     }
 
 
-class TestProjE(_ModelTestCase, unittest.TestCase):
+class TestProjE(cases.ModelTestCase):
     """Test the ProjE model."""
 
     model_cls = pykeen.models.ProjE
 
 
-class TestRESCAL(_ModelTestCase, unittest.TestCase):
+class TestRESCAL(cases.ModelTestCase):
     """Test the RESCAL model."""
 
     model_cls = pykeen.models.RESCAL
 
 
-class _TestRGCN(_ModelTestCase):
-    """Test the R-GCN model."""
-
-    model_cls = pykeen.models.RGCN
-    sampler = 'schlichtkrull'
-
-    def _check_constraints(self):
-        """Check model constraints.
-
-        Enriched embeddings have to be reset.
-        """
-        assert self.model.entity_representations.enriched_embeddings is None
-
-
-class TestRGCNBasis(_TestRGCN, unittest.TestCase):
+class TestRGCNBasis(cases.BaseRGCNTest):
     """Test the R-GCN model."""
 
     model_kwargs = {
@@ -777,7 +261,7 @@ class TestRGCNBasis(_TestRGCN, unittest.TestCase):
     num_constant_init = 2
 
 
-class TestRGCNBlock(_TestRGCN, unittest.TestCase):
+class TestRGCNBlock(cases.BaseRGCNTest):
     """Test the R-GCN model with block decomposition."""
 
     embedding_dim = 6
@@ -791,7 +275,7 @@ class TestRGCNBlock(_TestRGCN, unittest.TestCase):
     num_constant_init = 4
 
 
-class TestRotatE(_ModelTestCase, unittest.TestCase):
+class TestRotatE(cases.ModelTestCase):
     """Test the RotatE model."""
 
     model_cls = pykeen.models.RotatE
@@ -810,13 +294,13 @@ class TestRotatE(_ModelTestCase, unittest.TestCase):
         assert torch.allclose(relation_abs, torch.ones_like(relation_abs))
 
 
-class TestSimplE(_ModelTestCase, unittest.TestCase):
+class TestSimplE(cases.ModelTestCase):
     """Test the SimplE model."""
 
     model_cls = pykeen.models.SimplE
 
 
-class _BaseTestSE(_ModelTestCase, unittest.TestCase):
+class _BaseTestSE(cases.ModelTestCase):
     """Test the Structured Embedding model."""
 
     model_cls = pykeen.models.StructuredEmbedding
@@ -846,7 +330,7 @@ class TestSEHighMemory(_BaseTestSE):
     }
 
 
-class TestTransD(_DistanceModelTestCase, unittest.TestCase):
+class TestTransD(cases.DistanceModelTestCase):
     """Test the TransD model."""
 
     model_cls = pykeen.models.TransD
@@ -1005,7 +489,7 @@ class TestTransD(_DistanceModelTestCase, unittest.TestCase):
         assert (torch.norm(e_bot, dim=-1, p=2) <= 1.0 + 1.0e-06).all()
 
 
-class TestTransE(_DistanceModelTestCase, unittest.TestCase):
+class TestTransE(cases.DistanceModelTestCase):
     """Test the TransE model."""
 
     model_cls = pykeen.models.TransE
@@ -1019,7 +503,7 @@ class TestTransE(_DistanceModelTestCase, unittest.TestCase):
         assert torch.allclose(entity_norms, torch.ones_like(entity_norms))
 
 
-class TestTransH(_DistanceModelTestCase, unittest.TestCase):
+class TestTransH(cases.DistanceModelTestCase):
     """Test the TransH model."""
 
     model_cls = pykeen.models.TransH
@@ -1033,7 +517,7 @@ class TestTransH(_DistanceModelTestCase, unittest.TestCase):
         assert torch.allclose(entity_norms, torch.ones_like(entity_norms))
 
 
-class TestTransR(_DistanceModelTestCase, unittest.TestCase):
+class TestTransR(cases.DistanceModelTestCase):
     """Test the TransR model."""
 
     model_cls = pykeen.models.TransR
@@ -1087,7 +571,7 @@ class TestTransR(_DistanceModelTestCase, unittest.TestCase):
             assert all_in_bounds(emb(indices=None).norm(p=2, dim=-1), high=1., a_tol=1.0e-06)
 
 
-class TestTuckEr(_ModelTestCase, unittest.TestCase):
+class TestTuckEr(cases.ModelTestCase):
     """Test the TuckEr model."""
 
     model_cls = pykeen.models.TuckER
@@ -1098,7 +582,7 @@ class TestTuckEr(_ModelTestCase, unittest.TestCase):
     num_constant_init = 4
 
 
-class TestUM(_DistanceModelTestCase, unittest.TestCase):
+class TestUM(cases.DistanceModelTestCase):
     """Test the Unstructured Model."""
 
     model_cls = pykeen.models.UnstructuredModel
@@ -1123,7 +607,7 @@ class TestTesting(unittest.TestCase):
             for name, value in globals().items()
             if (
                 isinstance(value, type)
-                and issubclass(value, _ModelTestCase)
+                and issubclass(value, cases.ModelTestCase)
                 and not name.startswith('_')
                 and not issubclass(value.model_cls, MultimodalModel)
             )
